Extract SVGUnderlay and axis tick utils to own files
Moves SVGUnderlay component from SVGOverlay.tsx into its own SVGUnderlay.tsx file, and extracts defaultTickFormat/filterTicksByPixelDistance into axisTickUtils.ts. Re-exports SVGUnderlay from SVGOverlay for backward compatibility. Adds tests for top/right-only axis styling in both SVG layers and updates related test fixtures.
