+++ src/components/stream/SVGOverlay.tsx
@@ -28,7 +28,10 @@ import {
   resolveVerticalTickBaseline,
   tickPixelExtent
 } from "./svgOverlayUtils"
+import { defaultTickFormat, filterTicksByPixelDistance } from "./axisTickUtils"
 import { SVGChartTitle } from "./SVGChartTitle"
+
+export { SVGUnderlay } from "./SVGUnderlay"
 
 // ── Axis config ───────────────────────────────────────────────────────────
 //
@@ -136,214 +139,6 @@ interface SVGOverlayProps {
   linkedCrosshairSourceId?: string
 
   children?: ReactNode
-}
-
-// ── SVGUnderlay ─────────────────────────────────────────────────────────
-// Renders ONLY grid lines and axis baseline lines behind the canvas.
-
-interface SVGUnderlayProps {
-  width: number
-  height: number
-  totalWidth: number
-  totalHeight: number
-  margin: { top: number; right: number; bottom: number; left: number }
-  scales: StreamScales | null
-  showAxes?: boolean
-  axes?: AxisConfig[]
-  showGrid?: boolean
-  xFormat?: (d: number | Date | string, index?: number, allTicks?: number[]) => string | ReactNode
-  yFormat?: (d: number | Date | string) => string | ReactNode
-  axisExtent?: import("../charts/shared/axisExtent").AxisExtentMode
-}
-
-export function SVGUnderlay(props: SVGUnderlayProps) {
-  const {
-    width,
-    height,
-    totalWidth,
-    totalHeight,
-    margin,
-    scales,
-    showAxes,
-    axes,
-    showGrid,
-    xFormat,
-    yFormat,
-    axisExtent
-  } = props
-  const xTicks = useMemo(() => {
-    if (!scales) return []
-    const bottomAxis = axes?.find(a => a.orient === "bottom")
-    // Prefer the bottom axis config; when only a top axis is declared, inherit
-    // its extent/ticks so chart-level + per-axis overrides still apply.
-    const topAxis = axes?.find(a => a.orient === "top")
-    const xAxis = bottomAxis ?? topAxis
-    const extentMode = xAxis?.extent ?? axisExtent
-    const fmt = xAxis?.tickFormat || xFormat || defaultTickFormat
-    const maxFit = Math.max(2, Math.floor(width / 70))
-    const requested = xAxis?.ticks ?? 5
-    // Explicit `tickValues` bypasses both d3's "nice" generator and
-    // `axisExtent: "exact"` — the caller has hand-picked the positions
-    // and we honor them verbatim. Same contract as the ordinal frame's
-    // `rTickValues`. Pixel-distance filtering downstream still drops
-    // overlapping labels.
-    // Exact mode (without explicit values) honors the requested count
-    // rather than clamping it to `maxFit`; floor at 2 because
-    // `equidistantTicks` needs both endpoints.
-    const tickCount = extentMode === "exact" ? Math.max(2, requested) : Math.min(requested, maxFit)
-    const rawTicks = xAxis?.tickValues ?? ticksForMode(scales.x, tickCount, extentMode)
-    const rawValues = rawTicks.map(v => v.valueOf())
-    const candidates = rawTicks.map((v, i) => ({
-      value: v,
-      pixel: scales.x(v),
-      label: fmt(v, i, rawValues)
-    }))
-    // Estimate the widest label and use that as minimum spacing so labels
-    // (which are center-anchored) don't overlap.
-    const maxLabelWidth = candidates.reduce((max, c) => Math.max(max, typeof c.label === "string" ? c.label.length * 6.5 : typeof c.label === "number" ? String(c.label).length * 6.5 : 60), 0)
-    const minPx = Math.max(55, maxLabelWidth + 8)
-    return filterTicksByPixelDistance(candidates, minPx)
-  }, [scales, axes, xFormat, width, axisExtent])
-
-  const yTicks = useMemo(() => {
-    if (!scales) return []
-    const leftAxis = axes?.find(a => a.orient === "left")
-    const extentMode = leftAxis?.extent ?? axisExtent
-    const fmt = leftAxis?.tickFormat || yFormat || defaultTickFormat
-    const maxFit = Math.max(2, Math.floor(height / 30))
-    const requested = leftAxis?.ticks ?? 5
-    const tickCount = extentMode === "exact" ? Math.max(2, requested) : Math.min(requested, maxFit)
-    // Explicit `tickValues` wins over generated ticks — see xTicks comment.
-    const rawTicks = leftAxis?.tickValues ?? ticksForMode(scales.y, tickCount, extentMode)
-    const candidates = rawTicks.map(v => ({
-      value: v,
-      pixel: scales.y(v),
-      label: fmt(v)
-    }))
-    return filterTicksByPixelDistance(candidates, 22)
-  }, [scales, axes, yFormat, height, axisExtent])
-
-  const hasGrid = showGrid && scales
-  const hasBaselines = showAxes && scales
-
-  if (!hasGrid && !hasBaselines) return null
-
-  const bottomAxis = axes?.find(a => a.orient === "bottom")
-  const topAxis = axes?.find(a => a.orient === "top")
-  const leftAxis = axes?.find(a => a.orient === "left")
-  const rightAxis = axes?.find(a => a.orient === "right")
-  const xAxis = bottomAxis ?? topAxis
-  const yAxis = leftAxis ?? rightAxis
-  const xOrient = bottomAxis ? "bottom" : topAxis ? "top" : "bottom"
-  const yOrient = leftAxis ? "left" : rightAxis ? "right" : "left"
-  const xBaselineY = xOrient === "top" ? 0 : height
-  const yBaselineX = yOrient === "right" ? width : 0
-  const showXBaseline = hasBaselines && (xAxis ? xAxis.baseline !== false : true)
-  const showYBaseline = hasBaselines && (yAxis ? yAxis.baseline !== false : true)
-  const xJagged = xAxis?.jaggedBase || false
-  const yJagged = yAxis?.jaggedBase || false
-  const xAxisLine = resolveAxisLineStyle(xAxis?.axisStyle, { stroke: "var(--semiotic-border, #ccc)", strokeWidth: 1 })
-  const yAxisLine = resolveAxisLineStyle(yAxis?.axisStyle, { stroke: "var(--semiotic-border, #ccc)", strokeWidth: 1 })
-  const xGridLine = resolveAxisLineStyle((bottomAxis ?? topAxis)?.gridStyle, { stroke: "var(--semiotic-grid, #e0e0e0)", strokeWidth: 1 })
-  const yGridLine = resolveAxisLineStyle((leftAxis ?? rightAxis)?.gridStyle, { stroke: "var(--semiotic-grid, #e0e0e0)", strokeWidth: 1 })
-
-  return (
-    <svg
-      width={totalWidth}
-      height={totalHeight}
-      style={{
-        position: "absolute",
-        top: 0,
-        left: 0,
-        pointerEvents: "none"
-      }}
-    >
-      <g transform={`translate(${margin.left},${margin.top})`}>
-        {/* Grid lines */}
-        {hasGrid && (() => {
-          const bottomGridStyle = resolveGridDash((bottomAxis ?? topAxis)?.gridStyle)
-          const leftGridStyle = resolveGridDash((leftAxis ?? rightAxis)?.gridStyle)
-          const showXGrid = (bottomAxis ?? topAxis)?.grid !== false
-          const showYGrid = (leftAxis ?? rightAxis)?.grid !== false
-          return (
-          <g className="stream-grid">
-            {showXGrid && xTicks.map((tick, i) => (
-              <line
-                key={`xgrid-${i}`}
-                x1={tick.pixel}
-                y1={0}
-                x2={tick.pixel}
-                y2={height}
-                {...xGridLine}
-                strokeDasharray={bottomGridStyle ?? xGridLine.strokeDasharray}
-              />
-            ))}
-            {showYGrid && yTicks.map((tick, i) => (
-              <line
-                key={`ygrid-${i}`}
-                x1={0}
-                y1={tick.pixel}
-                x2={width}
-                y2={tick.pixel}
-                {...yGridLine}
-                strokeDasharray={leftGridStyle ?? yGridLine.strokeDasharray}
-              />
-            ))}
-          </g>
-          )
-        })()}
-
-        {/* Axis baselines */}
-        {showXBaseline && !xJagged && (
-          <line x1={0} y1={xBaselineY} x2={width} y2={xBaselineY} {...xAxisLine} />
-        )}
-        {xJagged && (
-          <path d={jaggedBaselinePath(xOrient, width, height)} fill="none" {...xAxisLine} />
-        )}
-        {showYBaseline && !yJagged && (
-          <line x1={yBaselineX} y1={0} x2={yBaselineX} y2={height} {...yAxisLine} />
-        )}
-        {yJagged && (
-          <path d={jaggedBaselinePath(yOrient, width, height)} fill="none" {...yAxisLine} />
-        )}
-      </g>
-    </svg>
-  )
-}
-
-
-function defaultTickFormat(v: string | number | Date, _index?: number, _allTicks?: number[]): string {
-  if (v instanceof Date) {
-    return `${v.toLocaleString("en", { month: "short" })} ${v.getDate()}`
-  }
-  if (typeof v === "number") return String(Math.round(v * 100) / 100)
-  return String(v)
-}
-
-/** Greedily filter ticks so consecutive labels are at least `minPx` apart.
- *  Always keeps the first and last tick. */
-// `value` is widened to `number | Date` so this helper accepts the
-// candidates produced for both linear scales and time-scale axes (the
-// latter emit Date instances; only `pixel` is used for distance math).
-function filterTicksByPixelDistance<T extends { value: number | Date; pixel: number; label: string | ReactNode }>(
-  ticks: T[],
-  minPx: number
-): T[] {
-  if (ticks.length <= 2) return ticks
-  const result = [ticks[0]]
-  for (let i = 1; i < ticks.length - 1; i++) {
-    if (Math.abs(ticks[i].pixel - result[result.length - 1].pixel) >= minPx) {
-      result.push(ticks[i])
-    }
-  }
-  const last = ticks[ticks.length - 1]
-  // Always keep the last tick (axis endpoint); if too close, replace the previous intermediate tick
-  if (Math.abs(last.pixel - result[result.length - 1].pixel) >= minPx) {
-    result.push(last)
-  } else {
-    result[result.length - 1] = last
-  }
-  return result
 }
 
 export function SVGOverlay(props: SVGOverlayProps) {
